requirements.txt
Flask==2.1.2
tensorflow==2.6.0
Pillow==8.2.0
opencv-python==4.5.3
